move test to right place
